Fix packaging setup
This commit makes a few small tweaks to the Python package so it's
installable and importable from a standalone package without the source
tree. There are a few things that are changed here, the first is the
namespace of the package is changed to somethign more unique, this could
be set via the package configuration in the setup.py, but it's easier to
just use the directory name. The second thing changed is a missing
__init__.py is added, this is needed to have python treat the directory
as a package and know it's importable. Third the setup.py is updated to
set the packages correctly and also add qiskit-terra as a requirement.